pyproject.toml
[project]
name = "predictive-hw2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kaggle>=2.0.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "scikit-learn>=1.8.0",
    "tensorflow[and-cuda]>=2.21.0",
    "torchinfo>=1.8.0",
    "tqdm>=4.67.3",
]
